\name{AEM-package}
\alias{AEM-package}
\alias{AEM}
\docType{package}
\title{
AEM - Asymmetric eigenvector map package
}
\description{
Package to construct asymmetric eigenvector maps (AEM).
}
\details{
\tabular{ll}{
Package: \tab AEM\cr
Type: \tab Package\cr
Version: \tab 0.2-3\cr
Date: \tab 2008-07-07\cr
License: \tab None... Copyleft \cr
}
This package construct spatial variables from the asymmetric eigenvector maps (AEM) framework. It was built following Blanchet et al. (In press).

}
\author{
F. Guillaume Blanchet

Maintainer: F. Guillaume Blanchet <gblanche@ualberta.ca>

}
\references{
Blanchet F.G., P. Legendre and Borcard D. 2008, Modelling directional spatial processes in ecological data. \emph{Ecological Modelling}, 215, 325-336.
}
