\name{AEM-package}
\alias{AEM-package}
\alias{AEM}
\docType{package}
\title{
AEM - Asymmetric eigenvector map package
}
\description{
Package to construct asymmetric eigenvector maps (AEM).
}
\details{
\tabular{ll}{
Package: \tab AEM\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2008-07-07\cr
License: \tab Unlimited \cr
}
This package construct spatial variables from the asymmetric eigenvector maps (AEM) framework. It was built following Blanchet et al. (2008).

}
\author{
F. Guillaume Blanchet

Maintainer: F. Guillaume Blanchet <guillaume.blanchet@helsinki.fi>

}
\references{
Blanchet F.G., P. Legendre and Borcard D. 2008, Modelling directional spatial processes in ecological data. \emph{Ecological Modelling}, 215, 325-336.
Blanchet, F. G., P. Legendre, R. Maranger, D. Monti, and P. Pepin. 2011. Modelling the effect of directional spatial ecological processes at different scales. \emph{Oecologia} 166, 357-368.
Legendre, P. and L. Legendre. 2012. Numerical ecology, 3rd English edition. Elsevier Science BV, Amsterdam. 
}
